fixes timestamp filtering of episode actions
The timestamp is being stored in the DB as an integer, but in a
field of type varchar(255). So the DB cannot be used to sort of filter,
because it would sort the integers alphabetically.

The workaround is to load all of the results into memory and then
throw away any that don't meet the optional `since=` filter.

fixes #39